database/requirements.txt
faker
psycopg2-binary
openai
redis[hiredis]
